+++ src/commands/CommandCommonInfo.ts
@@ -5,3 +5,4 @@ export const COPYTOCLIPBOARD_COMMAND = 'jsonHelper.copyToClipboard';
 export const MOVE_PRE_CMD = 'jsonHelper.moveToPreviousKey';
 export const MOVE_NEXT_CMD = 'jsonHelper.moveToNextKey';
 export const SHOW_NODES_QUICK_PICK_CMD = 'jsonHelper.showNodesInQuickPick';
+export const SHOW_JSON_OUTLINE = 'jsonHelper.jsonOutline';